refactor: refactor code smells
